+++ pkgs/dart_mcp_server/lib/src/utils/cli_utils.dart
@@ -10,6 +10,7 @@ import 'package:file/file.dart';
 import 'package:process/process.dart';
 import 'package:yaml/yaml.dart';
 
+import 'analytics.dart';
 import 'constants.dart';
 import 'sdk.dart';
 
@@ -159,7 +160,7 @@ Future<CallToolResult> runCommandInRoot(
         TextContent(text: 'Invalid root configuration: missing `root` key.'),
       ],
       isError: true,
-    );
+    )..failureReason ??= CallToolFailureReason.noRootGiven;
   }
 
   final root = knownRoots.firstWhereOrNull(
@@ -175,7 +176,7 @@ Future<CallToolResult> runCommandInRoot(
         ),
       ],
       isError: true,
-    );
+    )..failureReason ??= CallToolFailureReason.invalidRootPath;
   }
 
   final rootUri = Uri.parse(rootUriString);
@@ -189,7 +190,7 @@ Future<CallToolResult> runCommandInRoot(
         ),
       ],
       isError: true,
-    );
+    )..failureReason ??= CallToolFailureReason.invalidRootScheme;
   }
   final projectRoot = fileSystem.directory(rootUri);
 
@@ -213,7 +214,7 @@ Future<CallToolResult> runCommandInRoot(
         ),
       ],
       isError: true,
-    );
+    )..failureReason ??= CallToolFailureReason.invalidPath;
   }
   commandWithPaths.addAll(paths);
 
@@ -238,7 +239,7 @@ Future<CallToolResult> runCommandInRoot(
         ),
       ],
       isError: true,
-    );
+    )..failureReason ??= CallToolFailureReason.nonZeroExitCode;
   }
   return CallToolResult(
     content: [